requirements/local.txt
-r base.txt
django-debug-toolbar==1.2.2django-tastypie-swagger==0.1.2
